requirements.txt
fastapi==0.111.0
strawberry-graphql==0.227.4
pymongo==4.6.2
redis==5.0.3
firebase-admin==6.5.0
